+++ tsconfig.json
@@ -21,24 +21,24 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "src/pages/payment/index.js",
-    "src/pages/payment/Fail.jsx",
-    "src/pages/payment/Checkout.jsx",
-    "src/pages/payment/Success.jsx",
-    "src/pages/payment/server.js",
-    "src/pages/payment/payments.service.js",
-    "src/pages/payment/payments.router.js",
-    "src/pages/payment/payments.controller.js",
-    "src/pages/payment/PaymentForm.tsx",
+    "src/components/payment/index.js",
+    "src/components/payment/Fail.jsx",
+    "src/components/payment/Checkout.jsx",
+    "src/components/payment/Success.jsx",
+    "src/components/payment/server.js",
+    "src/components/payment/payments.service.js",
+    "src/components/payment/payments.router.js",
+    "src/components/payment/payments.controller.js",
+    "src/components/payment/PaymentForm.tsx",
     "src/pages/sandbox/index.js",
-    "src/components/Checkout.jsx",
-    "src/pages/payment/sandbox/Payment.jsx",
-    "src/pages/payment/sandbox/Success.jsx",
-    "src/pages/payment/sandbox/server.js",
-    "src/pages/payment/sandbox/payments.service.js",
-    "src/pages/payment/sandbox/payments.router.js",
-    "src/pages/payment/sandbox/payments.controller.js",
-    "src/pages/payment/sandbox/Fail.jsx"
+    "src/components/toss/Checkout.jsx",
+    "src/components/payment/sandbox/Payment.jsx",
+    "src/components/payment/sandbox/Success.jsx",
+    "src/components/payment/sandbox/server.js",
+    "src/components/payment/sandbox/payments.service.js",
+    "src/components/payment/sandbox/payments.router.js",
+    "src/components/payment/sandbox/payments.controller.js",
+    "src/components/payment/sandbox/Fail.jsx"
   ],
   "exclude": ["node_modules"]
 }
